apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-cloud-firestore==2.20.0
google-cloud-storage==2.19.0
google-cloud-pubsub==2.27.1
pydantic==2.10.3
firebase-admin==6.6.0
structlog==24.4.0
httpx==0.28.1
python-multipart==0.0.20
